bluetooth: mesh: remove deprecated legacy sensor API
- Removing kconfig option BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE and
removing legacy sensor API.
- Remove unit tests for legacy sensor APIs.
- Align uinttest light_ctrl with new sensor APIs.
