+++ include/bluetooth/mesh/sensor.h
@@ -36,7 +36,6 @@ extern "C" {
 /** String length for representing a single sensor channel. */
 #define BT_MESH_SENSOR_CH_STR_LEN 23
 
-#if !defined(CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE) || defined(__DOXYGEN__)
 /** @def BT_MESH_SENSOR_VALUE_IN_RANGE
  *
  *  @brief Returns whether or not encoded sensor value _value is in the range
@@ -49,15 +48,6 @@ extern "C" {
 #define BT_MESH_SENSOR_VALUE_IN_RANGE(_value, _start, _end) (                  \
 		(_value)->format->cb->compare((_value), (_start)) >= 0 &&      \
 		(_value)->format->cb->compare((_end), (_value)) >= 0)
-#else
-#define BT_MESH_SENSOR_VALUE_IN_RANGE(_value, _start, _end) (                  \
-		((_value)->val1 > (_start)->val1 ||                            \
-		 ((_value)->val1 == (_start)->val1 &&                          \
-		  (_value)->val2 >= (_start)->val2)) &&                        \
-		((_value)->val1 < (_end)->val1 ||                              \
-		 ((_value)->val1 == (_end)->val1 &&                            \
-		  (_value)->val2 <= (_end)->val2)))
-#endif
 
 /** Sensor sampling type.
  *
@@ -141,8 +131,6 @@ struct bt_mesh_sensor_type {
 
 struct bt_mesh_sensor;
 struct bt_mesh_sensor_srv;
-
-#if !defined(CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE) || defined(__DOXYGEN__)
 
 /** Status of conversion from @ref bt_mesh_sensor_value. */
 enum bt_mesh_sensor_value_status {
@@ -866,369 +854,6 @@ int bt_mesh_sensor_ch_to_str(const struct bt_mesh_sensor_value *ch, char *str,
  *  @return A string representing the sensor channel.
  */
 const char *bt_mesh_sensor_ch_str(const struct bt_mesh_sensor_value *ch);
-
-#else /* defined(CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE) */
-/** Sensor descriptor representing various static metadata for the sensor.
- */
-struct bt_mesh_sensor_descriptor {
-	/** Sensor measurement tolerance specification. */
-	struct {
-		/** Maximum positive measurement error (in percent).
-		 *  A tolerance of 0 should be interpreted as "unspecified".
-		 */
-		struct sensor_value positive;
-		/** Maximum negative measurement error (in percent).
-		 *  A tolerance of 0 should be interpreted as "unspecified".
-		 */
-		struct sensor_value negative;
-	} tolerance;
-	/** Sampling type for the sensor data. */
-	enum bt_mesh_sensor_sampling sampling_type;
-	/** Measurement period for the samples, if applicable. */
-	uint64_t period;
-	/** Update interval for the samples, if applicable. */
-	uint64_t update_interval;
-};
-
-/** Delta threshold type. */
-enum bt_mesh_sensor_delta {
-	/** Value based delta threshold.
-	 *  The delta threshold values are represented as absolute value
-	 *  changes.
-	 */
-	BT_MESH_SENSOR_DELTA_VALUE,
-	/** Percent based delta threshold.
-	 *  The delta threshold values are represented as percentages of their
-	 *  old value (resolution: 0.01 %).
-	 */
-	BT_MESH_SENSOR_DELTA_PERCENT,
-};
-
-/** Sensor thresholds for publishing. */
-struct bt_mesh_sensor_threshold {
-	/** Delta threshold values.
-	 *
-	 *  Denotes the minimal sensor value change that should cause the sensor
-	 *  to publish its value.
-	 */
-	struct {
-		/** Type of delta threshold */
-		enum bt_mesh_sensor_delta type;
-		/** Minimal delta for a positive change. */
-		struct sensor_value up;
-		/** Minimal delta for a negative change. */
-		struct sensor_value down;
-	} delta;
-	/** Range based threshold values.
-	 *
-	 *  Denotes the value range in which the sensor should be in fast
-	 *  cadence mode.
-	 */
-	struct {
-		/** Cadence when the sensor value is inside the range.
-		 *
-		 *  If the cadence is fast when the value is inside the range,
-		 *  it's normal when it's outside the range. If the cadence is
-		 *  normal when the value is inside the range, it's fast outside
-		 *  the range.
-		 */
-		enum bt_mesh_sensor_cadence cadence;
-		/** Lower boundary for the range based sensor cadence threshold.
-		 */
-		struct sensor_value low;
-		/** Upper boundary for the range based sensor cadence threshold.
-		 */
-		struct sensor_value high;
-	} range;
-};
-
-/** Sensor channel value format. */
-struct bt_mesh_sensor_format {
-	/** @brief Sensor channel value encode function.
-	 *
-	 *  @param[in]  format Pointer to the format structure.
-	 *  @param[in]  val    Sensor channel value to encode.
-	 *  @param[out] buf    Buffer to encode the value into.
-	 *
-	 *  @return 0 on success, or (negative) error code otherwise.
-	 */
-	int (*const encode)(const struct bt_mesh_sensor_format *format,
-			    const struct sensor_value *val,
-			    struct net_buf_simple *buf);
-
-	/** @brief Sensor channel value decode function.
-	 *
-	 *  @param[in]  format Pointer to the format structure.
-	 *  @param[in]  buf    Buffer to decode the value from.
-	 *  @param[out] val    Resulting sensor channel value.
-	 *
-	 *  @return 0 on success, or (negative) error code otherwise.
-	 */
-	int (*const decode)(const struct bt_mesh_sensor_format *format,
-			    struct net_buf_simple *buf,
-			    struct sensor_value *val);
-
-	/** User data pointer. Used internally by the sensor types. */
-	void *user_data;
-	/** Size of the encoded data in bytes. */
-	size_t size;
-
-#ifdef CONFIG_BT_MESH_SENSOR_LABELS
-	/** Pointer to the unit associated with this format. */
-	const struct bt_mesh_sensor_unit *unit;
-#endif
-};
-
-/** Single sensor setting. */
-struct bt_mesh_sensor_setting {
-	/** Sensor type of this setting. */
-	const struct bt_mesh_sensor_type *type;
-
-	/** @brief Getter for this sensor setting.
-	 *
-	 *  @note This handler is mandatory.
-	 *
-	 *  @param[in]  srv     Sensor server instance associated with this
-	 *                      setting.
-	 *  @param[in]  sensor  Sensor this setting belongs to.
-	 *  @param[in]  setting Pointer to this setting structure.
-	 *  @param[in]  ctx     Context parameters for the packet this call
-	 *                      originated from, or NULL if this call wasn't
-	 *                      triggered by a packet.
-	 *  @param[out] rsp     Response buffer for the setting value. Points to
-	 *                      an array with the number of channels specified
-	 *                      by the setting sensor type. All channels must be
-	 *                      filled.
-	 */
-	void (*get)(struct bt_mesh_sensor_srv *srv,
-		    struct bt_mesh_sensor *sensor,
-		    const struct bt_mesh_sensor_setting *setting,
-		    struct bt_mesh_msg_ctx *ctx,
-		    struct sensor_value *rsp);
-
-	/** @brief Setter for this sensor setting.
-	 *
-	 *  Should only be specified for writable sensor settings.
-	 *
-	 *  @param[in] srv     Sensor server instance associated with this
-	 *                     setting.
-	 *  @param[in] sensor  Sensor this setting belongs to.
-	 *  @param[in] setting Pointer to this setting structure.
-	 *  @param[in] ctx     Context parameters for the packet this call
-	 *                     originated from, or NULL if this call wasn't
-	 *                     triggered by a packet.
-	 *  @param[in] value   New setting value. Contains the number of
-	 *                     channels specified by the setting sensor type.
-	 *
-	 *  @return 0 on success, or (negative) error code otherwise.
-	 */
-	int (*set)(struct bt_mesh_sensor_srv *srv,
-		struct bt_mesh_sensor *sensor,
-		const struct bt_mesh_sensor_setting *setting,
-		struct bt_mesh_msg_ctx *ctx,
-		const struct sensor_value *value);
-};
-
-/** Single sensor series data column.
- *
- *  The series data columns represent a range for specific measurement values,
- *  inside which a set of sensor measurements were made. The range is
- *  interpreted as a half-open interval (i.e. start <= value < end).
- *
- *  @note Contrary to the Bluetooth Mesh specification, the column has an end
- *        value instead of a width, to match the conventional property format.
- *        This reduces implementation complexity for sensor series values that
- *        include the start and end (or min and max) of the measurement range,
- *        as the value of the column can be copied directly into the
- *        corresponding channels.
- */
-struct bt_mesh_sensor_column {
-	/** Start of the column (inclusive). */
-	struct sensor_value start;
-	/** End of the column (exclusive). */
-	struct sensor_value end;
-};
-
-/** Sensor series specification. */
-struct bt_mesh_sensor_series {
-	/** Pointer to the list of columns.
-	 *
-	 *  The columns may overlap, but the start value of each column must be
-	 *  unique. The list of columns do not have to cover the entire valid
-	 *  range, and values that don't fit in any of the columns should be
-	 *  ignored. If columns overlap, samples must be present in all columns
-	 *  they fall into. The columns may come in any order.
-	 *
-	 *  This list is not used for sensor types with one or two channels.
-	 */
-	const struct bt_mesh_sensor_column *columns;
-
-	/** Number of columns. */
-	uint32_t column_count;
-
-	/** @brief Getter for the series values.
-	 *
-	 *  Should return the historical data for the latest sensor readings in
-	 *  the given column.
-	 *
-	 *  @param[in]  srv          Sensor server associated with sensor instance.
-	 *  @param[in]  sensor       Sensor pointer.
-	 *  @param[in]  ctx          Message context pointer, or NULL if this call
-	 *                           didn't originate from a mesh message.
-	 *  @param[in]  column_index The index of the requested sensor column.
-	 *                           Index into the @c columns array for sensors
-	 *                           with more than two channels.
-	 *  @param[out] value        Sensor value response buffer. Holds the number
-	 *                           of channels indicated by the sensor type. All
-	 *                           channels must be filled.
-	 *
-	 *  @return 0 on success, or (negative) error code otherwise.
-	 */
-	int (*get)(struct bt_mesh_sensor_srv *srv,
-		struct bt_mesh_sensor *sensor,
-		struct bt_mesh_msg_ctx *ctx,
-		uint32_t column_index,
-		struct sensor_value *value);
-};
-
-/** Sensor instance. */
-struct bt_mesh_sensor {
-	/** Sensor type.
-	 *
-	 *  Must be one of the specification defined types listed in @ref
-	 *  bt_mesh_sensor_types.
-	 */
-	const struct bt_mesh_sensor_type *type;
-	/** Optional sensor descriptor. */
-	const struct bt_mesh_sensor_descriptor *descriptor;
-	/** Sensor settings access specification. */
-	const struct {
-		/** Static array of sensor settings */
-		const struct bt_mesh_sensor_setting *list;
-		/** Number of sensor settings. */
-		size_t count;
-	} settings;
-
-	/** Sensor series specification.
-	 *
-	 *  Only sensors who have a non-zero column-count and a defined
-	 *  series getter will accept series messages. Sensors with more than
-	 *  two channels also require a non-empty list of columns.
-	 */
-	const struct bt_mesh_sensor_series series;
-
-	/** @brief Getter function for the sensor value.
-	 *
-	 *  @param[in]  srv    Sensor server associated with sensor instance.
-	 *  @param[in]  sensor Sensor instance.
-	 *  @param[in]  ctx    Message context, or NULL if the call wasn't
-	 *                     triggered by a mesh message.
-	 *  @param[out] rsp    Value response buffer. Fits the number of
-	 *                     channels specified by the sensor type. All
-	 *                     channels must be filled.
-	 *
-	 *  @return 0 on success, or (negative) error code otherwise.
-	 */
-	int (*const get)(struct bt_mesh_sensor_srv *srv,
-			 struct bt_mesh_sensor *sensor,
-			 struct bt_mesh_msg_ctx *ctx,
-			 struct sensor_value *rsp);
-
-	/* Internal state, overwritten on init. Should only be written to by
-	 * internal modules.
-	 */
-	struct {
-		/** Sensor threshold specification. */
-		struct bt_mesh_sensor_threshold threshold;
-
-		/** Linked list node. */
-		sys_snode_t node;
-
-		/** The previously published sensor value. */
-		struct sensor_value prev;
-
-		/** Sequence number of the previous publication. */
-		uint16_t seq;
-
-		/** Minimum possible interval for fast cadence value publishing.
-		 *  The value is represented as 2 to the power of N milliseconds.
-		 *
-		 *  @see BT_MESH_SENSOR_INTERVAL_MAX
-		 */
-		uint8_t min_int;
-
-		/** Fast period divisor used when publishing with fast cadence.
-		 */
-		uint8_t pub_div : 4;
-
-		/** Flag indicating whether the sensor is in fast cadence mode.
-		 */
-		uint8_t fast_pub : 1;
-
-		/** Flag indicating whether the sensor cadence state has been configured. */
-		uint8_t configured : 1;
-	} state;
-};
-
-/** @brief Check if a value change breaks the delta threshold.
- *
- *  Sensors should publish their value if the measured sample is outside the
- *  delta threshold compared to the previously published value. This function
- *  checks the threshold and the previously published value for this sensor,
- *  and returns whether the sensor should publish its value.
- *
- *  @note Only single-channel sensors support cadence. Multi-channel sensors are
- *        always considered out of their threshold range, and will always return
- *        true from this function. Single-channel sensors that haven't been
- *        assigned a threshold will return true if the value is different.
- *
- *  @param[in] sensor    The sensor instance.
- *  @param[in] value     Sensor value.
- *
- *  @return true if the difference between the measurements exceeds the delta
- *          threshold, false otherwise.
- */
-bool bt_mesh_sensor_delta_threshold(const struct bt_mesh_sensor *sensor,
-				    const struct sensor_value *value);
-
-/** @brief Check whether a single channel sensor value lies within a column.
- *
- *  @param[in] value Value to check. Only the first channel is considered.
- *  @param[in] col   Sensor column.
- *
- *  @return true if the value belongs in the column, false otherwise.
- */
-bool bt_mesh_sensor_value_in_column(const struct sensor_value *value,
-				    const struct bt_mesh_sensor_column *col);
-
-/** @brief Get a human readable representation of a single sensor channel.
- *
- *  @param[in]  ch  Sensor channel to represent.
- *  @param[out] str String buffer to fill. Should be @ref
- *                  BT_MESH_SENSOR_CH_STR_LEN bytes long.
- *  @param[in]  len Length of @c str buffer.
- *
- *  @return Number of bytes that should have been written if @c str is
- *          sufficiently large.
- */
-static inline int bt_mesh_sensor_ch_to_str(const struct sensor_value *ch,
-					   char *str, size_t len)
-{
-	return snprintk(str, len, "%s%u.%06u",
-			((ch->val1 < 0 || ch->val2 < 0) ? "-" : ""),
-			abs(ch->val1), abs(ch->val2));
-}
-
-/** @brief Get a human readable representation of a single sensor channel.
- *
- *  @note This function is not thread safe.
- *
- *  @param[in] ch Sensor channel to represent.
- *
- *  @return A string representing the sensor channel.
- */
-const char *bt_mesh_sensor_ch_str(const struct sensor_value *ch);
-#endif /* !defined(CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE) */
 
 /** @brief Get the sensor type associated with the given Device Property ID.
  *
