+++ samples/bluetooth/mesh/light_ctrl/prj.conf
@@ -64,4 +64,3 @@ CONFIG_BT_MESH_LIGHT_CTRL_SRV_TIME_PROLONG=3
 CONFIG_BT_MESH_LIGHT_CTRL_SRV_RESUME_DELAY=30
 CONFIG_BT_MESH_SCENE_SRV=y
 CONFIG_BT_MESH_SENSOR_SRV=y
-CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE=n
