+++ samples/bluetooth/mesh/sensor_client/prj.conf
@@ -61,6 +61,5 @@ CONFIG_BT_MESH_SHELL=y
 
 # Bluetooth Mesh models
 CONFIG_BT_MESH_SENSOR_CLI=y
-CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE=n
 
 CONFIG_LOG_BACKEND_RTT=n
