+++ samples/bluetooth/mesh/sensor_server/prj.conf
@@ -64,7 +64,6 @@ CONFIG_BT_MESH_MODEL_KEY_COUNT=3
 # Bluetooth Mesh models
 CONFIG_BT_MESH_SENSOR_SRV=y
 CONFIG_BT_MESH_SENSOR_SRV_SENSORS_MAX=5
-CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE=n
 
 # Increase default delayable message config to avoid insuffisient resources
 # during normal sample operation
