+++ subsys/bluetooth/mesh/sensor.c
@@ -17,7 +17,6 @@ LOG_MODULE_REGISTER(bt_mesh_sensor);
 /** Scale away the sensor_value fraction, to allow integer math */
 #define SENSOR_MILL(_val) ((1000000LL * (_val)->val1) + (_val)->val2)
 
-#ifndef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
 /** Delta threshold type. */
 enum cadence_delta_type {
 	/** Value based delta threshold.
@@ -31,24 +30,11 @@ enum cadence_delta_type {
 	 */
 	CADENCE_DELTA_TYPE_PERCENT,
 };
-#endif
 
 static enum bt_mesh_sensor_cadence
 sensor_cadence(const struct bt_mesh_sensor_threshold *threshold,
-	       const sensor_value_type *curr)
+	       const struct bt_mesh_sensor_value *curr)
 {
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	int64_t high_mill = SENSOR_MILL(&threshold->range.high);
-	int64_t low_mill = SENSOR_MILL(&threshold->range.low);
-
-	if (high_mill == low_mill) {
-		return BT_MESH_SENSOR_CADENCE_NORMAL;
-	}
-
-	int64_t curr_mill = SENSOR_MILL(curr);
-	bool in_range = (curr_mill >= MIN(low_mill, high_mill) &&
-			 curr_mill <= MAX(low_mill, high_mill));
-#else
 	const struct bt_mesh_sensor_value *high, *low;
 	const struct bt_mesh_sensor_format *fmt = curr->format;
 
@@ -64,66 +50,10 @@ sensor_cadence(const struct bt_mesh_sensor_threshold *threshold,
 		return BT_MESH_SENSOR_CADENCE_NORMAL;
 	}
 
-	bool in_range = BT_MESH_SENSOR_VALUE_IN_RANGE(curr, low, high);
-#endif
-	return in_range ? threshold->range.cadence : !threshold->range.cadence;
+	return BT_MESH_SENSOR_VALUE_IN_RANGE(curr, low, high) ? threshold->range.cadence
+							      : !threshold->range.cadence;
 }
 
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-bool bt_mesh_sensor_delta_threshold(const struct bt_mesh_sensor *sensor,
-				    const struct sensor_value *curr)
-{
-	struct sensor_value delta = {
-		curr->val1 - sensor->state.prev.val1,
-		curr->val2 - sensor->state.prev.val2,
-	};
-	int64_t delta_mill = SENSOR_MILL(&delta);
-	int64_t thrsh_mill;
-
-	if (delta_mill < 0) {
-		delta_mill = -delta_mill;
-		thrsh_mill = SENSOR_MILL(&sensor->state.threshold.delta.down);
-	} else {
-		thrsh_mill = SENSOR_MILL(&sensor->state.threshold.delta.up);
-	}
-
-	/* If the threshold value is a perentage, we should calculate the actual
-	 * threshold value relative to the previous value.
-	 */
-	if (sensor->state.threshold.delta.type ==
-	    BT_MESH_SENSOR_DELTA_PERCENT) {
-		/* Store the Status Delta Trigger threshold in sensor_value first to avoid int64_t
-		 * overflow caused by multiplication of values converted by the SENSOR_MILL() macro.
-		 */
-		struct sensor_value thrsh_delta = delta_mill >= 0 ?
-			sensor->state.threshold.delta.up : sensor->state.threshold.delta.down;
-		struct sensor_value thrsh = {
-			.val1 = thrsh_delta.val1 * sensor->state.prev.val1,
-			.val2 = thrsh_delta.val2 * sensor->state.prev.val2,
-		};
-
-		thrsh_mill = llabs(SENSOR_MILL(&thrsh)) / 100LL;
-	}
-
-	if (IS_ENABLED(CONFIG_BT_MESH_MODEL_LOG_LEVEL_DBG)) {
-		char delta_str[BT_MESH_SENSOR_CH_STR_LEN];
-		char curr_str[BT_MESH_SENSOR_CH_STR_LEN];
-		char prev_str[BT_MESH_SENSOR_CH_STR_LEN];
-		char thrsh_str[BT_MESH_SENSOR_CH_STR_LEN];
-
-		strcpy(delta_str, bt_mesh_sensor_ch_str(&delta));
-		strcpy(curr_str, bt_mesh_sensor_ch_str(curr));
-		strcpy(prev_str, bt_mesh_sensor_ch_str(&sensor->state.prev));
-		strcpy(thrsh_str,
-		       delta_mill < 0 ? bt_mesh_sensor_ch_str(&sensor->state.threshold.delta.down) :
-					bt_mesh_sensor_ch_str(&sensor->state.threshold.delta.up));
-
-		LOG_DBG("Delta: %s (%s - %s) thrsh: %s", delta_str, curr_str, prev_str, thrsh_str);
-	}
-
-	return (delta_mill > thrsh_mill);
-}
-#else /* !defined(CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE) */
 bool bt_mesh_sensor_delta_threshold(const struct bt_mesh_sensor *sensor,
 				    const struct bt_mesh_sensor_value *curr)
 {
@@ -134,7 +64,6 @@ bool bt_mesh_sensor_delta_threshold(const struct bt_mesh_sensor *sensor,
 	return curr->format->cb->delta_check(curr, &sensor->state.prev,
 					     &sensor->state.threshold.deltas);
 }
-#endif /* defined(CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE) */
 
 void bt_mesh_sensor_cadence_set(struct bt_mesh_sensor *sensor,
 				enum bt_mesh_sensor_cadence cadence)
@@ -185,25 +114,6 @@ void sensor_status_id_decode(struct net_buf_simple *buf, uint8_t *len, uint16_t 
 	}
 }
 
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-static void tolerance_decode(uint16_t encoded, struct sensor_value *tolerance)
-{
-	uint32_t toll_mill = (encoded * 100ULL * 1000000ULL) / 4095ULL;
-
-	tolerance->val1 = toll_mill / 1000000ULL;
-	tolerance->val2 = toll_mill % 1000000ULL;
-}
-static uint16_t tolerance_encode(const struct sensor_value *tol)
-{
-	uint64_t tol_mill = 1000000ULL * tol->val1 + tol->val2;
-
-	if (tol_mill > (1000000ULL * 100ULL)) {
-		return 0;
-	}
-	return (tol_mill * 4095ULL + (1000000ULL * 50ULL)) / (1000000ULL * 100ULL);
-}
-#endif
-
 void sensor_descriptor_decode(struct net_buf_simple *buf,
 			struct bt_mesh_sensor_info *sensor)
 {
@@ -211,15 +121,8 @@ void sensor_descriptor_decode(struct net_buf_simple *buf,
 
 	sensor->id = net_buf_simple_pull_le16(buf);
 	tolerances = net_buf_simple_pull_le24(buf);
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	tolerance_decode(tolerances & BIT_MASK(12),
-			 &sensor->descriptor.tolerance.positive);
-	tolerance_decode(tolerances >> 12,
-			 &sensor->descriptor.tolerance.negative);
-#else
 	sensor->descriptor.tolerance.positive = tolerances & BIT_MASK(12);
 	sensor->descriptor.tolerance.negative = tolerances >> 12;
-#endif
 	sensor->descriptor.sampling_type = net_buf_simple_pull_u8(buf);
 	sensor->descriptor.period =
 		sensor_powtime_decode(net_buf_simple_pull_u8(buf));
@@ -236,13 +139,8 @@ void sensor_descriptor_encode(struct net_buf_simple *buf,
 	const struct bt_mesh_sensor_descriptor *d =
 		sensor->descriptor ? sensor->descriptor : &dummy;
 
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	uint16_t tol_pos = tolerance_encode(&d->tolerance.positive);
-	uint16_t tol_neg = tolerance_encode(&d->tolerance.negative);
-#else
 	uint16_t tol_pos = d->tolerance.positive;
 	uint16_t tol_neg = d->tolerance.negative;
-#endif
 
 	net_buf_simple_add_u8(buf, tol_pos & 0xff);
 	net_buf_simple_add_u8(buf,
@@ -256,7 +154,7 @@ void sensor_descriptor_encode(struct net_buf_simple *buf,
 
 int sensor_value_encode(struct net_buf_simple *buf,
 			const struct bt_mesh_sensor_type *type,
-			const sensor_value_type *values)
+			const struct bt_mesh_sensor_value *values)
 {
 	/* The API assumes that `values` array size is always CONFIG_BT_MESH_SENSOR_CHANNELS_MAX. */
 	__ASSERT_NO_MSG(type->channel_count <= CONFIG_BT_MESH_SENSOR_CHANNELS_MAX);
@@ -276,7 +174,7 @@ int sensor_value_encode(struct net_buf_simple *buf,
 
 int sensor_value_decode(struct net_buf_simple *buf,
 			const struct bt_mesh_sensor_type *type,
-			sensor_value_type *values)
+			struct bt_mesh_sensor_value *values)
 {
 	int err;
 
@@ -296,11 +194,8 @@ int sensor_value_decode(struct net_buf_simple *buf,
 
 int sensor_ch_encode(struct net_buf_simple *buf,
 		     const struct bt_mesh_sensor_format *format,
-		     const sensor_value_type *value)
+		     const struct bt_mesh_sensor_value *value)
 {
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	return format->encode(format, value, buf);
-#else
 	if (format != value->format) {
 		return -EINVAL;
 	}
@@ -311,16 +206,12 @@ int sensor_ch_encode(struct net_buf_simple *buf,
 
 	net_buf_simple_add_mem(buf, value->raw, value->format->size);
 	return 0;
-#endif
 }
 
 int sensor_ch_decode(struct net_buf_simple *buf,
 		     const struct bt_mesh_sensor_format *format,
-		     sensor_value_type *value)
+		     struct bt_mesh_sensor_value *value)
 {
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	return format->decode(format, buf, value);
-#else
 	if (buf->len < format->size) {
 		return -ENOMEM;
 	}
@@ -328,12 +219,11 @@ int sensor_ch_decode(struct net_buf_simple *buf,
 	memcpy(value->raw, net_buf_simple_pull_mem(buf, format->size),
 	       format->size);
 	return 0;
-#endif
 }
 
 int sensor_status_encode(struct net_buf_simple *buf,
 			 const struct bt_mesh_sensor *sensor,
-			 const sensor_value_type *values)
+			 const struct bt_mesh_sensor_value *values)
 {
 	const struct bt_mesh_sensor_type *type = sensor->type;
 	size_t size = 0;
@@ -367,7 +257,7 @@ int sensor_column_value_encode(struct net_buf_simple *buf,
 			       struct bt_mesh_msg_ctx *ctx,
 			       uint32_t column_index)
 {
-	sensor_value_type values[CONFIG_BT_MESH_SENSOR_CHANNELS_MAX];
+	struct bt_mesh_sensor_value values[CONFIG_BT_MESH_SENSOR_CHANNELS_MAX];
 	int err = sensor->series.get(srv, sensor, ctx, column_index, values);
 
 	if (err) {
@@ -392,18 +282,6 @@ int sensor_column_encode(struct net_buf_simple *buf,
 		return -ENOTSUP;
 	}
 
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	const uint64_t width_million =
-		(col->end.val1 - col->start.val1) * 1000000ULL +
-		(col->end.val2 - col->start.val2);
-	const struct sensor_value width = {
-		.val1 = width_million / 1000000ULL,
-		.val2 = width_million % 1000000ULL,
-	};
-
-	LOG_DBG("Column width: %s", bt_mesh_sensor_ch_str(&width));
-#endif
-
 	int err;
 
 	err = sensor_ch_encode(buf, col_format, &col->start);
@@ -411,12 +289,7 @@ int sensor_column_encode(struct net_buf_simple *buf,
 		return err;
 	}
 
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	/* The sensor columns are transmitted as start+width, not start+end: */
-	err = sensor_ch_encode(buf, col_format, &width);
-#else
 	err = sensor_ch_encode(buf, col_format, &col->width);
-#endif
 	if (err) {
 		return err;
 	}
@@ -427,7 +300,7 @@ int sensor_column_encode(struct net_buf_simple *buf,
 int sensor_column_decode(
 	struct net_buf_simple *buf, const struct bt_mesh_sensor_type *type,
 	struct bt_mesh_sensor_column *col,
-	sensor_value_type value[CONFIG_BT_MESH_SENSOR_CHANNELS_MAX])
+	struct bt_mesh_sensor_value value[CONFIG_BT_MESH_SENSOR_CHANNELS_MAX])
 {
 	const struct bt_mesh_sensor_format *col_format;
 	int err;
@@ -448,25 +321,10 @@ int sensor_column_decode(
 		return -ENOENT;
 	}
 
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	struct sensor_value width;
-
-	err = sensor_ch_decode(buf, col_format, &width);
-	if (err) {
-		return err;
-	}
-
-	uint64_t end_mill = (col->start.val1 + width.val1) * 1000000ULL +
-			 (col->start.val2 + width.val2);
-
-	col->end.val1 = end_mill / 1000000ULL;
-	col->end.val2 = end_mill % 1000000ULL;
-#else
 	err = sensor_ch_decode(buf, col_format, &col->width);
 	if (err) {
 		return err;
 	}
-#endif
 
 	return sensor_value_decode(buf, type, value);
 }
@@ -568,39 +426,22 @@ int sensor_cadence_encode(struct net_buf_simple *buf,
 			  uint8_t fast_period_div, uint8_t min_int,
 			  const struct bt_mesh_sensor_threshold *threshold)
 {
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	const struct bt_mesh_sensor_format *delta_format =
-		(threshold->delta.type == BT_MESH_SENSOR_DELTA_PERCENT) ?
-			&bt_mesh_sensor_format_percentage_delta_trigger :
-			sensor_type->channels[0].format;
-
-	enum bt_mesh_sensor_delta delta_type = !!threshold->delta.type;
-#else
 	const struct bt_mesh_sensor_format *delta_format = threshold->deltas.down.format;
 
 	enum cadence_delta_type delta_type =
 		delta_format == &bt_mesh_sensor_format_percentage_delta_trigger ?
 		CADENCE_DELTA_TYPE_PERCENT : CADENCE_DELTA_TYPE_VALUE;
-#endif
 	net_buf_simple_add_u8(buf, (delta_type << 7) |
 				   (BIT_MASK(7) & fast_period_div));
 
 	int err;
 
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	err = sensor_ch_encode(buf, delta_format, &threshold->delta.down);
-#else
 	err = sensor_ch_encode(buf, delta_format, &threshold->deltas.down);
-#endif
 	if (err) {
 		return err;
 	}
 
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	err = sensor_ch_encode(buf, delta_format, &threshold->delta.up);
-#else
 	err = sensor_ch_encode(buf, delta_format, &threshold->deltas.up);
-#endif
 	if (err) {
 		return err;
 	}
@@ -612,7 +453,7 @@ int sensor_cadence_encode(struct net_buf_simple *buf,
 	net_buf_simple_add_u8(buf, min_int);
 
 	/* Flip the order if the cadence is fast outside. */
-	const sensor_value_type *first, *second;
+	const struct bt_mesh_sensor_value *first, *second;
 
 	if (threshold->range.cadence == BT_MESH_SENSOR_CADENCE_FAST) {
 		first = &threshold->range.low;
@@ -636,45 +477,26 @@ int sensor_cadence_decode(struct net_buf_simple *buf,
 			  struct bt_mesh_sensor_threshold *threshold)
 {
 	const struct bt_mesh_sensor_format *delta_format;
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	enum bt_mesh_sensor_delta delta_type;
-#else
 	enum cadence_delta_type delta_type;
-#endif
 	uint8_t div_and_type;
 	int err;
 
 	div_and_type = net_buf_simple_pull_u8(buf);
 	delta_type = div_and_type >> 7;
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	threshold->delta.type = delta_type;
-	delta_format = (delta_type == BT_MESH_SENSOR_DELTA_PERCENT) ?
-			&bt_mesh_sensor_format_percentage_delta_trigger :
-			sensor_type->channels[0].format;
-#else
 	delta_format = (delta_type == CADENCE_DELTA_TYPE_PERCENT) ?
 			&bt_mesh_sensor_format_percentage_delta_trigger :
 			sensor_type->channels[0].format;
-#endif
 	*fast_period_div = div_and_type & BIT_MASK(7);
 	if (*fast_period_div > BT_MESH_SENSOR_PERIOD_DIV_MAX) {
 		return -EINVAL;
 	}
 
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	err = sensor_ch_decode(buf, delta_format, &threshold->delta.down);
-#else
 	err = sensor_ch_decode(buf, delta_format, &threshold->deltas.down);
-#endif
 	if (err) {
 		return err;
 	}
 
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	err = sensor_ch_decode(buf, delta_format, &threshold->delta.up);
-#else
 	err = sensor_ch_decode(buf, delta_format, &threshold->deltas.up);
-#endif
 	if (err) {
 		return err;
 	}
@@ -699,16 +521,6 @@ int sensor_cadence_decode(struct net_buf_simple *buf,
 	/* According to MshMDLv1.1: 4.1.3.6, if range.high is lower than range.low, the cadence is
 	 * fast outside the range. Swap the two in this case.
 	 */
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	if (threshold->range.high.val1 < threshold->range.low.val1 ||
-	    (threshold->range.high.val1 == threshold->range.low.val1 &&
-	     threshold->range.high.val2 < threshold->range.low.val2)) {
-		struct sensor_value temp;
-
-		temp = threshold->range.high;
-		threshold->range.high = threshold->range.low;
-		threshold->range.low = temp;
-#else
 	if (threshold->range.high.format->cb->compare(&threshold->range.low,
 						      &threshold->range.high) > 0) {
 		uint8_t temp[CONFIG_BT_MESH_SENSOR_CHANNEL_ENCODED_SIZE_MAX];
@@ -719,7 +531,6 @@ int sensor_cadence_decode(struct net_buf_simple *buf,
 		       CONFIG_BT_MESH_SENSOR_CHANNEL_ENCODED_SIZE_MAX);
 		memcpy(&threshold->range.low.raw, temp,
 		       CONFIG_BT_MESH_SENSOR_CHANNEL_ENCODED_SIZE_MAX);
-#endif
 		threshold->range.cadence = BT_MESH_SENSOR_CADENCE_NORMAL;
 	} else {
 		threshold->range.cadence = BT_MESH_SENSOR_CADENCE_FAST;
@@ -739,17 +550,6 @@ uint8_t sensor_pub_div_get(const struct bt_mesh_sensor *s, uint32_t base_period)
 	return div;
 }
 
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-bool bt_mesh_sensor_value_in_column(const struct sensor_value *value,
-				    const struct bt_mesh_sensor_column *col)
-{
-	return (value->val1 > col->start.val1 ||
-		(value->val1 == col->start.val1 &&
-		 value->val2 >= col->start.val2)) &&
-	       (value->val1 < col->end.val1 ||
-		(value->val1 == col->end.val1 && value->val2 <= col->end.val2));
-}
-#else
 bool bt_mesh_sensor_value_in_column(const struct bt_mesh_sensor_value *value,
 				    const struct bt_mesh_sensor_column *col)
 {
@@ -784,10 +584,9 @@ bool bt_mesh_sensor_value_in_column(const struct bt_mesh_sensor_value *value,
 
 	return valuef <= (startf + widthf);
 }
-#endif
 
 void sensor_cadence_update(struct bt_mesh_sensor *sensor,
-			   const sensor_value_type *value)
+			   const struct bt_mesh_sensor_value *value)
 {
 	if (!sensor->state.configured) {
 		/* Ignore any update before cadence is configured. */
@@ -815,11 +614,7 @@ void sensor_cadence_update(struct bt_mesh_sensor *sensor,
 	sensor->state.fast_pub = new;
 }
 
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-const char *bt_mesh_sensor_ch_str(const struct sensor_value *ch)
-#else
 const char *bt_mesh_sensor_ch_str(const struct bt_mesh_sensor_value *ch)
-#endif
 {
 	static char str[BT_MESH_SENSOR_CH_STR_LEN];
 
@@ -830,7 +625,6 @@ const char *bt_mesh_sensor_ch_str(const struct bt_mesh_sensor_value *ch)
 	return str;
 }
 
-#ifndef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
 int bt_mesh_sensor_ch_to_str(const struct bt_mesh_sensor_value *ch, char *str,
 			     size_t len)
 {
@@ -1009,4 +803,3 @@ int bt_mesh_sensor_value_from_special_status(
 	}
 	return format->cb->from_special_status(format, status, sensor_val);
 }
-#endif /* defined(CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE) */
