+++ subsys/bluetooth/mesh/sensor_cli.c
@@ -40,11 +40,7 @@ struct sensor_data_list_rsp {
 
 struct series_data_rsp {
 	struct bt_mesh_sensor_series_entry *entries;
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	const struct bt_mesh_sensor_column *col;
-#else
 	const struct bt_mesh_sensor_value *col_start;
-#endif
 	uint16_t id;
 	uint32_t count;
 };
@@ -156,7 +152,7 @@ static int handle_status(const struct bt_mesh_model *model, struct bt_mesh_msg_c
 			return -EMSGSIZE;
 		}
 
-		sensor_value_type value[CONFIG_BT_MESH_SENSOR_CHANNELS_MAX];
+		struct bt_mesh_sensor_value value[CONFIG_BT_MESH_SENSOR_CHANNELS_MAX];
 
 		err = sensor_value_decode(buf, type, value);
 		if (err) {
@@ -170,7 +166,7 @@ static int handle_status(const struct bt_mesh_model *model, struct bt_mesh_msg_c
 
 		if (rsp && count <= rsp->count) {
 			memcpy(rsp->sensors[count].value, value,
-			       sizeof(sensor_value_type) * type->channel_count);
+			       sizeof(struct bt_mesh_sensor_value) * type->channel_count);
 
 			rsp->sensors[count].type = type;
 			++count;
@@ -239,12 +235,7 @@ yield_ack:
 								  (void **)&rsp)) {
 		/* If column format exists verify Raw Value A */
 		if (col_format &&
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-			(rsp->col->start.val1 != entry.column.start.val1 ||
-			 rsp->col->start.val2 != entry.column.start.val2)) {
-#else
 			col_format->cb->compare(rsp->col_start, &entry.column.start) != 0) {
-#endif
 			return 0;
 		}
 
@@ -821,11 +812,7 @@ int bt_mesh_sensor_cli_setting_set(struct bt_mesh_sensor_cli *cli,
 				   struct bt_mesh_msg_ctx *ctx,
 				   const struct bt_mesh_sensor_type *sensor,
 				   const struct bt_mesh_sensor_type *setting,
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-				   const struct sensor_value *value,
-#else
 				   const struct bt_mesh_sensor_value *value,
-#endif
 				   struct bt_mesh_sensor_setting_status *rsp)
 {
 	int err;
@@ -874,11 +861,7 @@ int bt_mesh_sensor_cli_setting_set_unack(
 	struct bt_mesh_sensor_cli *cli, struct bt_mesh_msg_ctx *ctx,
 	const struct bt_mesh_sensor_type *sensor,
 	const struct bt_mesh_sensor_type *setting,
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	const struct sensor_value *value)
-#else
 	const struct bt_mesh_sensor_value *value)
-#endif
 {
 	int err;
 
@@ -938,11 +921,7 @@ int bt_mesh_sensor_cli_all_get(struct bt_mesh_sensor_cli *cli,
 int bt_mesh_sensor_cli_get(struct bt_mesh_sensor_cli *cli,
 			   struct bt_mesh_msg_ctx *ctx,
 			   const struct bt_mesh_sensor_type *sensor,
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-			   struct sensor_value *rsp)
-#else
 			   struct bt_mesh_sensor_value *rsp)
-#endif
 {
 	int err;
 
@@ -973,7 +952,7 @@ int bt_mesh_sensor_cli_get(struct bt_mesh_sensor_cli *cli,
 		return -ENODEV;
 	}
 
-	memcpy(rsp, sensor_data.value, sizeof(sensor_value_type) * sensor->channel_count);
+	memcpy(rsp, sensor_data.value, sizeof(struct bt_mesh_sensor_value) * sensor->channel_count);
 
 	return 0;
 }
@@ -981,11 +960,7 @@ int bt_mesh_sensor_cli_get(struct bt_mesh_sensor_cli *cli,
 int bt_mesh_sensor_cli_series_entry_get(
 	struct bt_mesh_sensor_cli *cli, struct bt_mesh_msg_ctx *ctx,
 	const struct bt_mesh_sensor_type *sensor,
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	const struct bt_mesh_sensor_column *column,
-#else
 	const union bt_mesh_sensor_column_key *column,
-#endif
 	struct bt_mesh_sensor_series_entry *rsp)
 {
 	const struct bt_mesh_sensor_format *col_format;
@@ -997,16 +972,6 @@ int bt_mesh_sensor_cli_series_entry_get(
 	net_buf_simple_add_le16(&msg, sensor->id);
 
 	col_format = bt_mesh_sensor_column_format_get(sensor);
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	if (col_format == NULL) {
-		net_buf_simple_add_le16(&msg, column->start.val1);
-	} else {
-		err = sensor_ch_encode(&msg, col_format, &column->start);
-		if (err) {
-			return err;
-		}
-	}
-#else
 	if (col_format == NULL) {
 		net_buf_simple_add_le16(&msg, column->index);
 	} else {
@@ -1015,17 +980,12 @@ int bt_mesh_sensor_cli_series_entry_get(
 			return err;
 		}
 	}
-#endif
 
 	struct series_data_rsp rsp_data = {
 		.entries = rsp,
 		.id = sensor->id,
 		.count = 1,
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-		.col = column,
-#else
 		.col_start = col_format ? &column->sensor_value : NULL,
-#endif
 	};
 
 	struct bt_mesh_msg_rsp_ctx rsp_ctx = {
@@ -1050,12 +1010,8 @@ int bt_mesh_sensor_cli_series_entry_get(
 int bt_mesh_sensor_cli_series_entries_get(
 	struct bt_mesh_sensor_cli *cli, struct bt_mesh_msg_ctx *ctx,
 	const struct bt_mesh_sensor_type *sensor,
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	const struct bt_mesh_sensor_column *range,
-#else
 	const union bt_mesh_sensor_column_key *range_start,
 	const union bt_mesh_sensor_column_key *range_end,
-#endif
 	struct bt_mesh_sensor_series_entry *rsp, uint32_t *count)
 {
 	int err;
@@ -1066,31 +1022,11 @@ int bt_mesh_sensor_cli_series_entries_get(
 
 	net_buf_simple_add_le16(&msg, sensor->id);
 
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	if (range) {
-#else
 	if (range_start && range_end) {
-#endif
 		const struct bt_mesh_sensor_format *col_format;
 
 		col_format = bt_mesh_sensor_column_format_get(sensor);
 
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-		if (!col_format) {
-			net_buf_simple_add_le16(&msg, range->start.val1);
-			net_buf_simple_add_le16(&msg, range->end.val1);
-		} else {
-			err = sensor_ch_encode(&msg, col_format, &range->start);
-			if (err) {
-				return err;
-			}
-
-			err = sensor_ch_encode(&msg, col_format, &range->end);
-			if (err) {
-				return err;
-			}
-		}
-#else
 		if (!col_format) {
 			net_buf_simple_add_le16(&msg, range_start->index);
 			net_buf_simple_add_le16(&msg, range_end->index);
@@ -1105,7 +1041,6 @@ int bt_mesh_sensor_cli_series_entries_get(
 				return err;
 			}
 		}
-#endif
 	}
 
 	struct series_data_rsp rsp_data = {
