+++ subsys/bluetooth/mesh/shell/shell_sensor_cli.c
@@ -17,16 +17,8 @@ static void descriptor_print(const struct shell *shell, int err,
 {
 	if (!err) {
 		shell_print(shell, "{");
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-		shell_fprintf(shell, SHELL_NORMAL, "\ttolerance: { positive: ");
-		shell_model_print_sensorval(shell, &rsp->tolerance.positive);
-		shell_fprintf(shell, SHELL_NORMAL, ", negative: ");
-		shell_model_print_sensorval(shell, &rsp->tolerance.negative);
-		shell_print(shell, " }");
-#else
 		shell_print(shell, "\ttolerance: { positive: %d, negative: %d }",
 			    rsp->tolerance.positive, rsp->tolerance.negative);
-#endif
 		shell_print(shell, "\tsampling type: %d", rsp->sampling_type);
 		shell_print(shell, "\tperiod: %lld", rsp->period);
 		shell_print(shell, "\tupdate interval: %lld", rsp->update_interval);
@@ -91,13 +83,6 @@ static void cadence_print(const struct shell *shell, int err,
 	if (!err) {
 		shell_print(shell, "fast period div: %d\nmin interval: %d", rsp->fast_period_div,
 			    rsp->min_int);
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-		shell_fprintf(shell, SHELL_NORMAL, "delta threshold: { type: %d, up: ",
-			      rsp->threshold.delta.type);
-		shell_model_print_sensorval(shell, &rsp->threshold.delta.up);
-		shell_fprintf(shell, SHELL_NORMAL, ", down: ");
-		shell_model_print_sensorval(shell, &rsp->threshold.delta.down);
-#else
 		int delta_type = rsp->threshold.deltas.up.format ==
 			&bt_mesh_sensor_format_percentage_delta_trigger;
 		shell_fprintf(shell, SHELL_NORMAL, "delta threshold: { type: %d, up: ",
@@ -105,7 +90,6 @@ static void cadence_print(const struct shell *shell, int err,
 		shell_model_print_sensorval(shell, &rsp->threshold.deltas.up);
 		shell_fprintf(shell, SHELL_NORMAL, ", down: ");
 		shell_model_print_sensorval(shell, &rsp->threshold.deltas.down);
-#endif
 		shell_fprintf(shell, SHELL_NORMAL,
 			      " }\nfast cadence range: { cadence inside: %d, lower boundary: ",
 			      rsp->threshold.range.cadence);
@@ -162,12 +146,7 @@ static int cadence_set(const struct shell *shell, size_t argc, char *argv[], boo
 		.fast_period_div = shell_strtoul(argv[2], 0, &err),
 		.min_int = shell_strtoul(argv[3], 0, &err),
 		.threshold = {
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-			.delta = {
-				.type = shell_strtol(argv[4], 0, &err),
-#else
 			.deltas = {
-#endif
 				.up = shell_model_strtosensorval(delta_format, argv[5], &err),
 				.down = shell_model_strtosensorval(delta_format, argv[6], &err)
 			},
@@ -253,7 +232,7 @@ static int cmd_settings_get(const struct shell *shell, size_t argc, char *argv[]
 	return err;
 }
 
-static void values_print(const struct shell *shell, int err, sensor_value_type *values,
+static void values_print(const struct shell *shell, int err, struct bt_mesh_sensor_value *values,
 			 const struct bt_mesh_sensor_type *sensor_type)
 {
 	if (!err) {
@@ -315,8 +294,8 @@ static int setting_set(const struct shell *shell, size_t argc, char *argv[], boo
 		return -ENOENT;
 	}
 
-	sensor_value_type value = shell_model_strtosensorval(setting_type->channels[0].format,
-							     argv[3], &err);
+	struct bt_mesh_sensor_value value =
+		shell_model_strtosensorval(setting_type->channels[0].format, argv[3], &err);
 
 	if (err) {
 		shell_warn(shell, "Unable to parse input string arg");
@@ -391,7 +370,7 @@ static int cmd_get(const struct shell *shell, size_t argc, char *argv[])
 		return err;
 	}
 
-	sensor_value_type rsp[CONFIG_BT_MESH_SENSOR_CHANNELS_MAX];
+	struct bt_mesh_sensor_value rsp[CONFIG_BT_MESH_SENSOR_CHANNELS_MAX];
 	const struct bt_mesh_sensor_type *sensor_type = bt_mesh_sensor_type_get(sensor_id);
 
 	if (sensor_type == NULL) {
@@ -411,13 +390,8 @@ static void series_entry_print(const struct shell *shell, int err,
 	if (!err) {
 		shell_fprintf(shell, SHELL_NORMAL, "[");
 		shell_model_print_sensorval(shell, &entry->column.start);
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-		shell_fprintf(shell, SHELL_NORMAL, " to ");
-		shell_model_print_sensorval(shell, &entry->column.end);
-#else
 		shell_fprintf(shell, SHELL_NORMAL, ", width ");
 		shell_model_print_sensorval(shell, &entry->column.width);
-#endif
 		shell_fprintf(shell, SHELL_NORMAL, "]: ");
 		values_print(shell, err, entry->value, sensor_type);
 	}
@@ -433,11 +407,6 @@ static int cmd_series_entry_get(const struct shell *shell, size_t argc, char *ar
 		return -ENOENT;
 	}
 
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-	struct bt_mesh_sensor_column column = {
-		.start = shell_model_strtosensorval(NULL, argv[2], &err)
-	};
-#else
 	union bt_mesh_sensor_column_key column;
 
 	if (sensor_type->channel_count == 3) {
@@ -446,7 +415,6 @@ static int cmd_series_entry_get(const struct shell *shell, size_t argc, char *ar
 	} else {
 		column.index = shell_strtoul(argv[2], 0, &err);
 	}
-#endif
 
 	if (err) {
 		shell_warn(shell, "Unable to parse input string arg");
@@ -503,12 +471,6 @@ static int cmd_series_entries_get(const struct shell *shell, size_t argc, char *
 	uint32_t count = CONFIG_BT_MESH_SHELL_SENSOR_CLI_MAX_COLUMNS;
 
 	if (argc == 4) {
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-		struct bt_mesh_sensor_column range = {
-			.start = shell_model_strtosensorval(NULL, argv[2], &err),
-			.end = shell_model_strtosensorval(NULL, argv[3], &err)
-		};
-#else
 		union bt_mesh_sensor_column_key range_start, range_end;
 
 		if (sensor_type->channel_count == 3) {
@@ -520,26 +482,17 @@ static int cmd_series_entries_get(const struct shell *shell, size_t argc, char *
 			range_start.index = shell_strtoul(argv[2], 0, &err);
 			range_end.index = shell_strtoul(argv[3], 0, &err);
 		}
-#endif
 
 		if (err) {
 			shell_warn(shell, "Unable to parse input string arg");
 			return err;
 		}
 
-#ifdef CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE
-		err = bt_mesh_sensor_cli_series_entries_get(cli, NULL, sensor_type, &range, rsp,
-							    &count);
-	} else if (argc == 2) {
-		err = bt_mesh_sensor_cli_series_entries_get(cli, NULL, sensor_type, NULL, rsp,
-							    &count);
-#else
 		err = bt_mesh_sensor_cli_series_entries_get(cli, NULL, sensor_type, &range_start,
 							    &range_end, rsp, &count);
 	} else if (argc == 2) {
 		err = bt_mesh_sensor_cli_series_entries_get(cli, NULL, sensor_type, NULL, NULL, rsp,
 							    &count);
-#endif
 	} else {
 		return -ENOEXEC;
 	}
