+++ subsys/bluetooth/mesh/shell/shell_utils.h
@@ -8,10 +8,10 @@
 #include <zephyr/shell/shell.h>
 #include "../sensor.h"
 
-sensor_value_type shell_model_strtosensorval(const struct bt_mesh_sensor_format *format,
-					     const char *str, int *err);
+struct bt_mesh_sensor_value shell_model_strtosensorval(const struct bt_mesh_sensor_format *format,
+						       const char *str, int *err);
 
-void shell_model_print_sensorval(const struct shell *shell, sensor_value_type *value);
+void shell_model_print_sensorval(const struct shell *shell, struct bt_mesh_sensor_value *value);
 
 double shell_model_strtodbl(const char *str, int *err);
 
@@ -20,7 +20,7 @@ bool shell_model_first_get(uint16_t id, const struct bt_mesh_model **mod);
 bool shell_vnd_model_first_get(uint16_t cid, uint16_t id, const struct bt_mesh_model **mod);
 
 int shell_model_instance_set(const struct shell *shell, const struct bt_mesh_model **mod,
-			      uint16_t mod_id, uint8_t elem_idx);
+			     uint16_t mod_id, uint8_t elem_idx);
 
 int shell_vnd_model_instance_set(const struct shell *shell, const struct bt_mesh_model **mod,
 				 uint16_t mod_id, uint16_t cid, uint8_t elem_idx);
