+++ tests/bluetooth/tester/prj.conf
@@ -82,8 +82,6 @@ CONFIG_BT_MESH_LIGHT_CTRL_SRV_TIME_FADE_STANDBY_AUTO=0
 CONFIG_BT_MESH_LIGHT_CTRL_SRV_TIME_FADE_STANDBY_MANUAL=0
 CONFIG_BT_MESH_LIGHT_CTRL_SRV_TIME_PROLONG=0
 
-CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE=n
-
 CONFIG_SETTINGS=y
 CONFIG_FLASH=y
 CONFIG_FLASH_MAP=y
