+++ tests/subsys/bluetooth/mesh/light_ctrl/CMakeLists.txt
@@ -58,8 +58,8 @@ target_compile_options(app
   -DCONFIG_BT_MESH_LIGHT_CTRL_SRV_REG_LUX_PROLONG=200
   -DCONFIG_BT_MESH_LIGHT_CTRL_SRV_REG_LUX_STANDBY=100
   -DCONFIG_BT_MESH_LIGHT_CTRL_AMB_LIGHT_LEVEL_TIMEOUT=20
-  -DCONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE=1
   -DCONFIG_BT_MESH_SENSOR_CHANNELS_MAX=1
+  -DCONFIG_BT_MESH_SENSOR_CHANNEL_ENCODED_SIZE_MAX=4
   -DCONFIG_BT_MESH_USES_TINYCRYPT
 )
