+++ tests/subsys/bluetooth/mesh/models/prj.conf
@@ -67,4 +67,3 @@ CONFIG_BT_MESH_LIGHT_HSL_CLI=y
 CONFIG_BT_MESH_SCHEDULER_CLI=y
 CONFIG_DM_MODULE=y
 CONFIG_BT_MESH_DM_CLI=y
-CONFIG_BT_MESH_SENSOR_USE_LEGACY_SENSOR_VALUE=n
