+++ tests/subsys/bluetooth/mesh/sensor_subsys/prj.conf
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2021-2023 Nordic Semiconductor ASA
+# Copyright (c) 2024 Nordic Semiconductor ASA
 #
 # SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
 #
