--- tests/subsys/bluetooth/mesh/sensor_subsys_new/prj.conf
@@ -1,10 +1,0 @@
-#
-# Copyright (c) 2024 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-# Ztest configuration
-CONFIG_ZTEST=y
-CONFIG_CBPRINTF_FP_SUPPORT=y
-CONFIG_NET_BUF=y
